add note to self on the redunancy of dout mask and enable
