+++ pueo/turfio/pueo_hsalign.py
@@ -1,266 +1,152 @@
-# Generic high-speed align and biterr module.
-from ..common.bf import bf
+# Reworked generic high-speed align and biterr module.
+# This is the base module. The TURFIO subclasses this out
+# in to pueo_hsdout and pueo_hscout.
+# Note that some of the functions here are actually 'common'
+# between the two - for instance, the train enable (which
+# controls the output, not the inputs) sets on both.
 from ..common.dev_submod import dev_submod
-
-from enum import Enum
 import time
 
-# Support module for PUEO high speed serial alignment.
-#
-# We support 2 types of highspeed align: 8 bit and 32 bit
-# Each has their own specific train value and max number
-# of bitslips.
-# Right now this is JUST for the TURFIO<->SURF interface.
-# The UltraScales work a bit different, but we'll probably
-# refactor this so that the methods are common at some point.
-# We also might merge in the RXCLK alignment as well,
-# making a base class or something. Who knows.
-
-# sigh, just force the RXCLK alignment here for the
-# caligns as an option.
 class PueoHSAlign(dev_submod):
-    map = { 'CTLRESET' : 0x0,
-            'IDELAY' : 0x4,
-            'BITERR' : 0x8,
-            'BITSLP' : 0xC,
-            'SYSERR' : 0x1C }
-
-    class BitWidth(Enum):
-        BITWIDTH_8 = 8
-        BITWIDTH_32 = 32
-        
-
-    # OK, the map is so much more complicated now.
-    # We have 2 things we need to derive: number of bitslips
-    # and the capture phase. But we'll 'encode' that
-    # in the top bit, so 7/6/5/4 need capture phase 1,
-    # and 3/2/1/0 need capture phase 0.
-    # You can only determine this map after a reset of the ISERDES
-    # and when you're in capture phase 0!
-    # Once you bitslip you don't really 'know' where you are!
+    # maps are here for convenience
     
-    # This map isn't exactly number of bitslips: bit 3 tells you
-    # if you need to flip DOUT's capture phase.
-    # In other words, you do the number of bitslips in this
-    # number & 0x3, and if this number & 0x4, you set dout
-    # capture phase to 1.
-    #
-    # D4 -> A9 -> 53 -> A6 (3 bitslips needed, dout capture phase 1)
-    # A9 -> 53 -> A6    (2 bitslips needed, dout capture phase 1)
-    # 53 -> A6 (1 bitslip needed, dout capture phase 1)
-    # A6 (0 bitslips needed, dout capture phase 1)
-    # NOTE: I'M ACTUALLY NOT SURE ABOUT THESE ONES
-    # 4D (3 bitslips needed, dout capture phase 0)
-    # 9A (2 bitslips needed, dout capture phase 0)
-    # 35 (1 bitslip needed, dout capture phase 0)
-    # A6 (0 bitslips needed, dout capture phase 0)
+    # the top bit here is actually the dout
+    # capture phase.
     BW8_MAP = { 0xD4 : 7,
                 0xA9 : 6,
                 0x53 : 5,
                 0xA6 : 4,
                 0x4D : 3,
                 0x9A : 2,
-                0x35 : 1,
+                0x35 : 1
                 0x6A : 0 }
 
-    # LOCK_REQ is either a lock request (if lockable) or enable
-    # TRAINEN enables training on the associated output interface
-    class CtlReset(Enum):
-        ISERDES_RST = 2
-        OSERDES_RST = 4
-        LOCK_RST = 3
-        LOCK_REQ = 8
-        LOCKED = 9
-        TRAINEN = 10
+    # The 32-bit map only needs the 4 bit-rotated values
+    # and their nybble rotations. We don't need to distinguish
+    # between the nybble rotations because we lock to
+    # the proper pattern.
+    BW32_MAP = { 0xa55a6996 : 0,
+                 0x55a6996a : 0,
+                 0x5a69965a : 0,
+                 0xa6996a55 : 0,
+                 0x6996a55a : 0,
+                 0x996a55a6 : 0,
+                 0x96a55a69 : 0,
+                 0x6a55a699 : 0,
 
-    # train value used in high-speed alignment
-    trainVal = { BitWidth.BITWIDTH_8 : 0x6A,
-                 BitWidth.BITWIDTH_32 : 0xA55A6996 }
+                 0x52AD34CB : 1,
+                 0x2AD34CB5 : 1,
+                 0xAD34CB52 : 1,
+                 0xD34CB52A : 1,
+                 0x34CB52AD : 1,
+                 0x4CB52AD3 : 1,
+                 0xCB52AD34 : 1,
+                 0xB52AD34C : 1,
+                 
+                 0xA9569A65 : 2,
+                 0x9569A65A : 2,
+                 0x569A65A9 : 2,
+                 0x69A65A95 : 2,
+                 0x9A65A956 : 2,
+                 0xA65A9569 : 2,
+                 0x65A9569A : 2,
+                 0x5A9569A6 : 2,
+                 
+                 0xD4AB4D32 : 3,
+                 0x4AB4D32D : 3,
+                 0xAB4D32D4 : 3,
+                 0xB4D32D4A : 3,
+                 0x4D32D4AB : 3,
+                 0xD32D4AB4 : 3,
+                 0x32D4AB4D : 3,
+                 0x2D4AB4D3 : 3 }
+                
+    # the map is for convenient lookup but I don't
+    # use them in the module functions for speed
+    map = { 'CTLRESET' : 0x0,
+            'IDELAY' : 0x4,
+            'BITERR' : 0x8,
+            'BITSLP' : 0xC }
 
-    # maximum number of bitslips (SERDES width)
-    maxSlips = { BitWidth.BITWIDTH_8 : 8,
-                 BitWidth.BITWIDTH_32 : 4 }
-    
-        
-    # static function
-    # This returns either None (incorrect eye value)
-    # or the number of bit-slips required to match up.
-    # Training pattern is 0xA55A6996. The bit-slipped
-    # versions of that are:
-    # 0xA55A6996 (0 bitslips needed)
-    # 0x52AD34CB (1 bitslip  needed)
-    # 0xA9569A65 (2 bitslips needed)
-    # 0xD4AB4D32 (3 bitslips needed)
-    # GODDAMNIT JUST MAKE A LOOKUP TABLE
-    @classmethod
-    def check_eye(cls, eye_val, bw=32,
-                  trainValue=trainVal[BitWidth.BITWIDTH_32]):   
-        testVal = int(eye_val)
-        # just hardcode this check for now!!!
-        if (bw == 8):
-            return cls.BW8_MAP[testVal] if testVal in cls.BW8_MAP else None
-        def rightRotate(n, d):
-            return (n>>d)|(n<<(bw-d)) & (2 ** bw - 1)
-        for i in range(bw):
-            if testVal == rightRotate(trainValue, i):
-                return i
-        return None    
-
-    # bw is an Enum of the bit widths
-    # maxTaps is the number of taps available in the IDELAY
-    # eyeInTaps is the width of the data eye in units of taps
-    # (78.125 ps tap width, 2 ns eye width = 25.6
     def __init__(self, dev, base,
-                 lockable=False,
-                 bw=BitWidth.BITWIDTH_32,
-                 maxTaps=32,
-                 eyeInTaps=26):
-        self.bw = bw
+                 bit_width = 32,
+                 max_idelay_taps = 63,
+                 eye_tap_width = 26,
+                 train_map = None):
+        if train_map is None:
+            raise Exception("must provide a training map of values to slips")
+
+        self.bit_width = bit_width
         self.lockable = lockable
-        self.maxTaps = maxTaps
-        self.eyeInTaps = eyeInTaps
+        self.max_taps = max_idelay_taps
+        self.eye_tap_width = eye_tap_width
+        self.train_map = train_map
+
         super().__init__(dev, base)
 
-    @staticmethod
-    def edges_to_eyes(scan, edges, maxWidth=32, eyeWidth=26):
-        eyes = {}
-        for edge in edges:
-            edgeVal = int((edge[0] + edge[1])/2)
-            eyeCenter = edgeVal - (eyeWidth//2)
-            if eyeCenter > 0 and scan[eyeCenter] is not None:
-                eyes[scan[eyeCenter][1]] = eyeCenter
-        lastEyeCenter = edgeVal + (eyeWidth//2)
-        if lastEyeCenter < maxWidth and scan[lastEyeCenter] is not None:
-            eyes[scan[lastEyeCenter][1]] = lastEyeCenter
-        return eyes
-        
-    # Find the edges of an eyescan
-    @staticmethod
-    def process_eyescan_edge(scan, width=32, verbose=False, allEdges=False):
+    # These are the scan processing functions.
+    @classmethod
+    def process_eyescan_edge(cls, scan, verbose=False):
+        """
+        Takes an eyescan - a tuple of (# errors, offset value)
+        and returns tuples of the value borders.
+        """
         edges = []
         start = None
         stop = None
-        curBitno = None
-        for i in range(width):
+        cur_bitno = None
+        for i in range(len(scan)):
             val = scan[i]
-            if val[0] == 0 and val[1] is not None:
-                # no errors, data OK
-                if curBitno is None:
-                    curBitno = val[1]
+            if val[0] and val[1] is not None:
+                # data OK
+                if cur_bitno is None:
+                    cur_bitno = val[1]
                     start = i
-                elif val[1] != curBitno:
+                elif val[1] != cur_bitno:
                     stop = i
+                    edge = (start, stop)
                     if verbose:
-                        print("edge at (", start, ",", stop, ")")
-                    if not allEdges:
-                        return (start, stop)
-                    edges.append((start,stop))
+                        print("edge at", eye)
+                    edges.append(eye)
                     start = None
                     stop = None
-                    curBitno = None
+                    cur_bitno = None
                 else:
                     start = i
         if len(edges) == 0:
             return None
         return edges
 
-    # RXCLK scan method
-    @staticmethod
-    def process_eyescan(scan, width=32, wrap=False):
-        scanShift = 0
-        if wrap:
-            if scan[0] == 0:
-                scanShift = next((i for i, x in enumerate(scan[::-1]) if x), None)
-                # roll the scan, then adjust back
-                scan = scan[-1*scanShift:] + scan[:-1*scanShift]
-                
-        # We start off by assuming we're not in an eye.
-        in_eye = False
-        eye_start = 0
-        eyes = []
-        for i in range(width):
-            if scan[i] == 0 and not in_eye:
-                eye_start = i
-                in_eye = True
-            elif scan[i] > 0 and in_eye:                
-                eye = [ int(eye_start+(i-eye_start)/2), i-eye_start ]
-                # now adjust it
-                if scanShift != 0:
-                    eyePos = eye[0]
-                    eyePos -= scanShift
-                    if eyePos < 0:
-                        eyePos += width
-                    eye = [ eyePos , i-eye_start ]
-                eyes.append(eye)
-                in_eye = False
-        # we exited the loop without finding the end of the eye
-        if in_eye:
-            eye = [ int(eye_start+(width-eye_start)/2), width-eye_start ]
-            eyes.append( eye )
-
+    @classmethod
+    def process_eyescan_eyes(cls, scan, eye_width=26, verbose=False):
+        edges = cls.process_eyescan_edge(scan, verbose=verbose)
+        if edges is None:
+            return edges
+        max_width = len(scan)
+        eyes = {}
+        for edge in edges:
+            edge_val = int((edge[0] + edge[1])/2)
+            eye_center = edge_val - (eye_width//2)
+            if eye_center > 0 and scan[eye_center] is not None:
+                eyes[scan[eye_center][1]] = eye_center
+        last_eye_center = edge_val + (eye_width//2)
+        if last_eye_center < max_width and scan[last_eye_center] is not None:
+            eyes[scan[last_eye_center][1]] = last_eye_center
         return eyes
-    
-    # convenience function for setting delay
-    # Might end up subclassing the HSAlign for the UltraScale/7 series differences
-    def set_delay(self, delayVal):
-        # if we have more than 32 taps, we're cascaded
-        # and our cascade sleaze is that when we jump into the second
-        # IDELAY we need to add 1.
-        if self.maxTaps > 32:
-            delayVal = delayVal+1 if delayVal & 32 else delayVal
-        self.write(self.map['IDELAY'], delayVal)
-    
-    def eyescan(self, slptime=0.01, getBitno=True):
-        sc = []
-        for i in range(self.maxTaps):
-            self.set_delay(i)
-            time.sleep(slptime)
-            biterr = self.read(self.map['BITERR'])
-            if getBitno:
-                if biterr == 0:
-                    testVal = self.read(self.map['BITSLP'])
-                    nbits = self.check_eye(testVal, self.bw.value, self.trainVal[self.bw])
-                    if nbits != None:
-                        sc.append((biterr, nbits % self.maxSlips[self.bw]))
-                    else:
-                        sc.append((biterr, None))
-                else:
-                    sc.append((biterr, None))
-            else:
-                sc.append(biterr)
-        return sc
-
-    # rxclk eyescan for the rxclk-capable
-    def eyescan_rxclk(self, period=1024):
-        slptime = period*8E-9
-        sc = []
-        self.rxclk_phase = 0
-        self.write(0x1C, period)
-        for i in range(448):
-            self.rxclk_phase = i
-            time.sleep(slptime)
-            sc.append(self.read(0x1C))
-        return sc
 
     @property
-    def rxclk_phase(self):
-        return self.read(0)>>16
-    
-    @rxclk_phase.setter
-    def rxclk_phase(self, value):
-        rv = self.read(0) & 0xFFFF
-        rv |= (value & 0xFFFF) << 16
-        self.write(0, rv)        
+    def idelay(self):
+        r = self.read(4)
+        if self.max_taps > 32:
+            if r & 32:
+                r = (r & 63) - 1
+        return r
 
-    @property
-    def dout_mask(self):
-        return (self.read(0)>>15) & 0x1
-
-    @dout_mask.setter
-    def dout_mask(self, value):
-        rv = self.read(0) & 0xFFFF7FFF
-        rv |= 0x8000 if value else 0
-        self.write(0, rv)
+    @idelay.setter:
+    def idelay(self, value):
+        if self.max_taps > 32:
+            value = value + 1 if value & 32 else value
+        self.write(0x4, value)
 
     @property
     def iserdes_reset(self):
@@ -273,120 +159,63 @@ class PueoHSAlign(dev_submod):
         self.write(0, rv)
 
     @property
-    def dout_capture_phase(self):
-        return (self.read(0) >> 7) & 0x1
+    def train_enable(self):
+        return (self.read(0) >> 10) & 0x1
 
-    @dout_capture_phase.setter
-    def dout_capture_phase(self, value):
-        rv = self.read(0) & 0xFFFFFF7F
-        rv |= 0x80 if value else 0
+    @train_enable.setter
+    def train_enable(self, value):
+        rv = self.read(0) & 0xfffffbff
+        rv |= 0x400 if value else 0
         self.write(0, rv)
-        
-    # enable training on output interface
-    def trainEnable(self, onOff):
-        rv = bf(self.read(self.map['CTLRESET']))
-        rv[self.CtlReset.TRAINEN.value] = int(onOff)
-        self.write(self.map['CTLRESET'], int(rv))
-    
-    # enable or lock.
-    def enable(self, onOff, verbose=False):
-        if onOff:
-            # enable
-            rv = bf(self.read(self.map['CTLRESET']))
-            rv[self.CtlReset.LOCK_REQ.value] = 1
-            self.write(self.map['CTLRESET'], int(rv))
-            rv = bf(self.read(self.map['CTLRESET']))
-            # locked is either a copy of LOCK_REQ or the status
-            if rv[self.CtlReset.LOCKED.value]:
-                if verbose:
-                    print("Interface enabled and running.")
-                return True
-            else:
-                raise Exception("Interface did not enable")
-        else:
-            # disable
-            if self.lockable:
-                rv = bf(self.read(self.map['CTLRESET']))
-                rv[self.CtlReset.LOCK_RST.value] = 1
-                self.write(self.map['CTLRESET'], int(rv))
-            else:
-                rv = bf(self.read(self.map['CTLRESET']))
-                rv[self.CtlReset.LOCK_REQ.value] = 0
-                self.write(self.map['CTLRESET'], int(rv))
-            return True
 
-    # OK: we now return a dict of eyes.
-    # That way in startup, we can get dicts for each of the
-    # devices, and then use Python's bitwise-and functions
-    # for lists to find a common eye between all of them.
-    # So no more individual sync offsets!
-    def find_alignment(self, doReset=True, verbose=False):
-        if (doReset):
+    def eyescan(self, slptime=0.01, get_bitno=True, cntclks=131072):
+        self.write(0x8, cntclks)
+        sc = []
+        for i in range(self.max_taps):
+            self.idelay = i
+            time.sleep(slptime)
+            biterr = self.read(0x8)
+            if get_bitno:
+                if biterr == 0:
+                    test = self.read(0xC)
+                    nb = train_map[test] if test in train_map else None
+                    sc.append((biterr, nb))
+                else:
+                    sc.append((biterr, None))
+            else:
+                sc.append(biterr)
+        return sc
+        
+    def find_alignment(self, do_reset=True, verbose=False):
+        """
+        Find the eyes in a link. Returns a dictionary
+        of eye values to tap number.
+        Raises exception if no eyes found.
+        """
+        if do_reset:
             self.iserdes_reset = 1
             self.iserdes_reset = 0
 
-        delayVal = None
-        slipFix = None
-
-        # set the biterr time and select our input
-        self.write(self.map['BITERR'], 131072)
+        delay = None
+        slip = None
+    
         sc = self.eyescan()
-        # Find ALL edges
-        edge = self.process_eyescan_edge(sc, width=self.maxTaps, allEdges=True)
-        if edge is None:
-            raise IOError("Cannot find eye transition: check training status")
+        eyes = self.process_eyescan_eyes(sc,
+                                         eye_width=self.eye_tap_width,
+                                         verbose=verbose)
+        if eyes is None:
+            raise IOError("Cannot find eyes: check training status")
 
-        return self.edges_to_eyes(sc, edge,
-                                  maxWidth=self.maxTaps,
-                                  eyeWidth=self.eyeInTaps)
+        return eyes
 
-    # now this function takes a tuple of (eye, bitslips)
-    # This is becoming a super mess, we really need to
-    # subclass out things for DOUT vs COUT.
     def apply_alignment(self, eye, verbose=False):
-        self.set_delay(eye[0])
-        slipFix = eye[1]
-        if verbose:
-            print("Performing", slipFix, "bit slips")
-        for i in range(slipFix):
-            self.write(self.map['BITSLP'], 1)
-
-        val = self.read(self.map['BITSLP'])
-        checkVal = self.check_eye(val, self.bw.value, self.trainVal[self.bw])
-        if checkVal % self.maxSlips[self.bw]:
-            raise IOError("Alignment procedure failed: still need %d slips??" % (checkVal % self.maxSlips[self.bw]))
-        else:
-            if verbose:
-                print("Alignment succeeded")
-            return True
-    
-    # Alignment method for RXCLK (for the TURF CALIGN)
-    def align_rxclk(self, verbose=False):
-        if verbose:
-            print("Scanning RXCLK->SYSCLK transition.")
-        rxsc = self.eyescan_rxclk()
-        eyes = self.process_eyescan(rxsc, 448, wrap=True)
-        bestEye = None
-        for eye in eyes:
-            if bestEye is not None:
-                if verbose:
-                    print(f'Second RXCLK->SYSCLK eye found at {eye[0]}, possible glitch')
-                if eye[1] > bestEye[1]:
-                    if verbose:
-                        print(f'Eye at {eye[0]} has width {eye[1]}, better than {bestEye[1]}')
-                    bestEye = eye
-                else:
-                    if verbose:
-                        print(f'Eye at {eye[0]} has width {eye[1]}, worse than {bestEye[1]}, skipping')
-            else:
-                if verbose:
-                    print(f'First eye at {eye[0]} width {eye[1]}')
-                bestEye = eye
-        if bestEye is None:
-            raise IOError("No valid RXCLK->SYSCLK eye found!!")
-        if verbose:
-            print(f'Using eye at {eye[0]}')
-        self.rxclk_phase = eye[0]
-        return bestEye[0]
-    
-                        
+        """
+        Applies an eye alignment to a link.
+        Eye is a tuple of (delay, #bitslips)
+        Note that this function is overridden
+        with 8-bit stuff bc we also have dout
+        capture phase to deal with.
+        """
+        
+        
+
